pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "backend"
version = "0.0.1"
description = "Wansport Computer Vision"
authors = [
  { name = "Gianluca Calò" },
]

readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    
]
requires-python = "~=3.12"
dependencies = [
    "fastapi[standard]>=0.118.0",
    "oracledb>=3.3.0",
    "pydantic>=2.11.10",
    "python-dotenv>=1.1.1",
]
